server/requirements.txt

# Flask Requirements
flask
flask_cors
flask_login
flask_bcrypt

# Development Requirements
# logging

# Mongo DB Requirements
pymongo

# Other Requirements
python-dotenv

